bot/cogs/challenge.py
import discord
from discord.ext import commands

class ChallengeCog(commands.Cog):
    """Challenge system commands (placeholder)"""
    
    def __init__(self, bot):
        self.bot = bot
    
    @commands.command(name='challenge')
    async def create_challenge(self, ctx):
        """Create a challenge"""
        embed = discord.Embed(
            title="⚔️ Challenge System",
            description="Challenge system coming soon!",
            color=discord.Color.orange()
        )
        await ctx.send(embed=embed)

async def setup(bot):
    await bot.add_cog(ChallengeCog(bot))