requirements.txt
discord.py>=2.3.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pandas>=2.0.0
asyncio-mqtt>=0.16.0